Add troubleshooting guide to README
As a new user of the react-hot-boilerplate it took me a while to understand why some files were hot-loadable and others were not, though in retrospect it's pretty obvious. As I strayed from the boilerplate, things mysteriously stopped working. The link to the troubleshooting guide should help, since it already has a section on that topic.